requirements.txt
# =========================
# Core scientific stack
# =========================
# Python 3.8 compatible pins
numpy==1.24.4
scipy==1.10.1
pandas==2.0.3
matplotlib==3.7.5
Pillow>=10.1.0,<11
opencv-python>=4.10,<4.12
scikit-image==0.21.0
scikit-learn==1.3.2
PyYAML==6.0.2
tqdm==4.67.1
imageio>=2.34,<2.38
packaging==24.2

# =========================
# PyTorch stack
# =========================
# If you need CUDA, install a matching torch/torchvision build from:
# https://pytorch.org/get-started/locally/
torch>=2.1.0,<2.6
torchvision>=0.16.0,<0.21
torchmetrics>=1.3,<1.6
torchdiffeq==0.2.5
pytorch-ignite>=0.4.13,<0.6
pytorch-fid>=0.3.0,<0.4

# =========================
# Flow Matching / CFM / PnP
# =========================
deepinv==0.2.0
POT==0.9.4
lpips==0.1.4
piq==0.8.0
ml-collections==0.1.1
gdown==5.2.0
ninja==1.11.1.1

# =========================
# U-Net helpers
# =========================
einops>=0.8,<0.9
ema-pytorch>=0.5.2,<0.6
accelerate>=0.29,<1.8

# =========================
# Demo app
# =========================
streamlit>=1.33,<1.45
